requirements.txt
accelerate>=0.27.0
albumentations
bitsandbytes
clip @ git+https://github.com/openai/CLIP.git
clip_retrieval
Cython
diffusers>=0.30.2
einops
hdf5storage
huggingface_hub
ipyfilechooser
ipython
ipywidgets
matplotlib
natsort
numpy
opencv-python
packaging
pandas
Pillow
PyYAML
regex
Requests
safetensors
scipy
scikit-image
tensorly
torch
torchaudio
torchvision
tqdm
transformers>=4.28.0
wandb
xformers
